+++ sdk/containerregistry/azure-containers-containerregistry/CHANGELOG.md
@@ -17,8 +17,9 @@
   - Removed `Collection<ManifestMediaType> mediaTypes` parameter from `downloadManifestWithResponse` method on blob clients.
   - Renamed `downloadStream(String, WritableByteChannel, Context`) to `downloadStreamWithResponse`. It also now returns `Response<Void>`.
   - Renamed `ContainerRegistryContentClientBuilder.repository` method to `repositoryName`.
-- Renamed `GetManifestResult.getMediaType` and `UploadManifestOptions.getMediaType` to `getManifestMediaType`
+- Renamed `GetManifestResult.getMediaType` and `UploadManifestOptions.getMediaType` to `getManifestMediaType`.
 - Removed `GetManifestResult.asOciManifest` - use `GetManifestResult.getManifest().toObject(OciImageManifest.class)` instead.
+- Renamed `OciImageManifest.getConfig` and `setConfig` to `getConfiguration` and `setConfiguration`.
 
 ### Bugs Fixed